utils: Move `hash_password()` and `check_password()` to shared module
These functions can then be used by other pieces of code that need to
deal with password hashing, such as the ApiKey model and factory.
